refactoring: moved configuration to a separate files
Templates was moved to:
- template/argument.json
- template/context.js

Configuration was separated from context and moved to:
- template/config.json

Functions parseSection, parseUsage, parseExamples now exports to template/context.js and was moved to:
- parseSection.js

Functions setArgument, setDesctription was moved to:
- argumentsFill.js

Function parseSection was renamed to findSection

Closes #39
