+++ src/handle.js
@@ -1,44 +1,16 @@
 'use strict';
 
-const argumentTemplate = {
-    shortName: null,
-    longName: null,
-    isFlag: true,
-    default: false,
-    description: null,
-};
+const argumentTemplate = require('./template/argument.json');
+const context = require('./template/context.js');
 
-const handle = (help) => {
-    const context = {
-        section: [
-            {
-                name: 'Options:',
-                data: [],
-                required: true,
-                func: parseOptions,
-            },
-            {
-                name: 'Usage:',
-                data: [],
-                required: false,
-                func: parseUsage,
-            },
-            {
-                name: 'Examples:',
-                data: [],
-                required: false,
-                func: parseExamples,
-            },
-        ],
-        options: [],
-        delimiter: '',
-    };
-
+const handle = (help, config) => {
     try {
-        context.section.forEach((section) => {
-            section.data = parseSection(section.name, help);
+        Object.keys(context.section).forEach((sectionName) => {
+            let section = context.section[sectionName];
+            section.data = findSection(section.name, help);
             if (section.data) {
-                section.data.forEach((data) => section.func(data, context));
+                section.data.forEach((data) =>
+                    section.func(data, context, argumentTemplate));
             } else if (section.required) {
                 const sectionNotFound
                     = new Error(`Required section ${section.name} not found`);
@@ -51,42 +23,7 @@ const handle = (help) => {
     }
 };
 
-const parseOptions = (section, context) => {
-    try {
-        splitSection(section).forEach((option) => {
-            if (option.indexOf('-') === 0) {
-                let argument = Object.assign({}, argumentTemplate);
-
-                option.trim().split(/\s\s+/, 2).map((section, index) => {
-                    index === 0 ? setArgument(section, argument)
-                        : setDescription(section, argument);
-                });
-                context.options.push(argument);
-            }
-        });
-    } catch (error) {
-        throw error;
-    }
-};
-
-const parseUsage = (section, context) => {
-    let argument = Object.assign({}, argumentTemplate);
-    argument.longName = '';
-    context.options.push(section.match(/file|path/i) ? argument : {});
-};
-
-const parseExamples = (section, context) => {
-    const delimiterDefault = '=';
-    const delimitersTemplate = [' ', '='];
-    delimitersTemplate.forEach((delimiter) => {
-        const regularExp = new RegExp(`-[^ \t\n]+${delimiter}[^ \t\n-]`);
-        section.match(regularExp) ? context.delimiter = delimiter : '';
-    });
-    context.delimiter =
-        context.delimiter ? context.delimiter : delimiterDefault;
-};
-
-const parseSection = (sectionName, help) => {
+const findSection = (sectionName, help) => {
     try {
         const regularExp =
             new RegExp(`[^\n]*${sectionName}[^\n]*\n?(?:[ \t].*?(?:\n|$))*`);
@@ -95,65 +32,6 @@ const parseSection = (sectionName, help) => {
     } catch (error) {
         throw error;
     }
-};
-
-const splitSection = (section) => {
-    try {
-        let result = [];
-        const strings = section.split('\n').slice(1);
-
-        let j = 0;
-
-        strings.forEach((str) => {
-            str = str.trim();
-
-            if (str.indexOf('-') === 0) {
-                result.push(str);
-                j++;
-            } else if (j > 0) {
-                result[j - 1] = result[j - 1] + str;
-            }
-        });
-
-        return result;
-    } catch (error) {
-        throw error;
-    }
-};
-
-const setArgument = (section, argument) => {
-    const argumentPrefix = /^-*/i;
-
-    section = removeExtraCharacters(section);
-    section.split(/\s+/).map((arg) => {
-        switch (argumentPrefix.exec(arg).toString()) {
-            case '--':
-                argument.longName = arg;
-                break;
-            case '-':
-                argument.shortName = arg;
-                break;
-            default:
-                argument.isFlag = false;
-                break;
-        }
-    });
-};
-
-const setDescription = (section, argument) => {
-    argument.description = section;
-    // TODO: add other patterns for default value parsing
-    const defaultValues = section.match(/\[default: (.*?)\]/i);
-    if (defaultValues) {
-        argument.defaultValue = defaultValues[1];
-        argument.isFlag = false;
-    }
-};
-
-const removeExtraCharacters = (args) => {
-    args = args.replace(/=/g, ' ');
-    args = args.replace(/,/g, ' ');
-    return args;
 };
 
 module.exports = handle;
