+++ src/parser.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const parser = (help) => {
+const parser = (help, config) => {
     const handle = require('./handle.js');
     const templatizer = require('./templatizer.js');
 
-    return templatizer(handle(help));
+    return templatizer(handle(help, config));
 };
 
 module.exports = parser;
