+++ src/template/argument.json
@@ -1,0 +1,7 @@
+{
+    "shortName": null,
+    "longName": null,
+    "isFlag": true,
+    "default": null,
+    "description": null
+}
