+++ src/template/config.json
@@ -1,0 +1,14 @@
+{
+    "section": {
+        "options": {
+            "name": "Options:"
+        },
+        "usage": {
+            "name": "Usage:"
+        },
+        "examples": {
+            "name": "Examples:"
+        }
+    },
+    "delimiter": "="
+}
